# app/Http/Controllers/CompteController.php
<?php

namespace App\Http\Controllers;

use App\Http\Requests\CompteStoreRequest;
use App\Http\Requests\CompteBloquerRequest;
use App\Http\Requests\CompteUpdateRequest;
use App\Http\Resources\CompteResource;
use App\Http\Resources\CompteResourceCollection;
use App\Services\ClientService;
use App\Services\CompteService;
use Illuminate\Http\Request;
use App\Traits\ApiResponser;
use Symfony\Component\HttpFoundation\Response;

/**
 * @OA\Info(
 *     title="API de Gestion des Comptes",
 *     version="1.0.0",
 *     description="API pour la gestion des comptes bancaires"
 * )
 *
 * @OA\Server(
 * url="https://jeeri.onrender.com/tinkin",
 * description="Serveur de production"
 * )
 * @OA\Server(
 * url="http://localhost:8000/tinkin",
 * description="Serveur local"
 * )
 *
 * @OA\Tag(
 *     name="Comptes",
 *     description="Gestion des comptes bancaires et de leurs transactions"
 * )
 */
class CompteController extends Controller
{
    use ApiResponser;

    protected CompteService $compteService;
    protected ClientService $clientService;

    public function __construct(CompteService $compteService, ClientService $clientService)
    {
        $this->compteService = $compteService;
        $this->clientService = $clientService;
    }

    /**
     * Liste paginée des comptes
     *
     * @OA\Get(
     *     path="/v1/comptes",
     *     tags={"Comptes"},
     *     summary="Liste paginée des comptes",
     *     description="Récupère une liste paginée de tous les comptes avec possibilité de filtrage",
     *     @OA\Parameter(name="page", in="query", @OA\Schema(type="integer", default=1)),
     *     @OA\Parameter(name="limit", in="query", @OA\Schema(type="integer", default=10)),
     *     @OA\Parameter(name="type", in="query", @OA\Schema(type="string", enum={"cheque", "epargne", "courant"})),
     *     @OA\Parameter(name="statut", in="query", @OA\Schema(type="string")),
     *     @OA\Parameter(name="search", in="query", @OA\Schema(type="string")),
     *     @OA\Parameter(name="sort", in="query", @OA\Schema(type="string", default="id")),
     *     @OA\Parameter(name="order", in="query", @OA\Schema(type="string", enum={"asc", "desc"}, default="desc")),
     *     @OA\Response(
     *         response=200,
     *         description="Liste des comptes récupérée avec succès",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="success"),
     *             @OA\Property(property="http_code", type="integer", example=200),
     *             @OA\Property(property="message", type="string", example="Liste des comptes récupérée avec succès"),
     *             @OA\Property(
     *                 property="data",
     *                 type="array",
     *                 @OA\Items(ref="#/components/schemas/Compte")
     *             ),
     *             @OA\Property(property="pagination", ref="#/components/schemas/Pagination")
     *         )
     *     )
     * )
     */
    public function index(Request $request)
    {
        $filters = $request->all();
        $page = (int) $request->get('page', 1);
        $limit = (int) $request->get('limit', 10);

        $comptes = $this->compteService->getAll($filters, $page, $limit);

        return $this->successResponse(
            $comptes,
            "Liste des comptes récupérée avec succès"
        );
    }

    /**
     * Liste des comptes non archivés
     *
     * @OA\Get(
     *     path="/v1/comptes/non-archives",
     *     tags={"Comptes"},
     *     summary="Liste des comptes non archivés",
     *     description="Récupère tous les comptes non archivés avec possibilité de filtrage",
     *     @OA\Parameter(name="page", in="query", @OA\Schema(type="integer", default=1)),
     *     @OA\Parameter(name="limit", in="query", @OA\Schema(type="integer", default=10)),
     *     @OA\Parameter(name="type", in="query", @OA\Schema(type="string", enum={"cheque", "epargne", "courant"})),
     *     @OA\Parameter(name="statut", in="query", @OA\Schema(type="string")),
     *     @OA\Parameter(name="search", in="query", @OA\Schema(type="string")),
     *     @OA\Parameter(name="sort", in="query", @OA\Schema(type="string", default="id")),
     *     @OA\Parameter(name="order", in="query", @OA\Schema(type="string", enum={"asc","desc"}, default="desc")),
     *     @OA\Response(
     *         response=200,
     *         description="Liste des comptes non archivés récupérée avec succès",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="success"),
     *             @OA\Property(property="http_code", type="integer", example=200),
     *             @OA\Property(property="message", type="string", example="Liste des comptes non archives récupérée avec succès"),
     *             @OA\Property(
     *                 property="data",
     *                 type="array",
     *                 @OA\Items(ref="#/components/schemas/Compte")
     *             ),
     *             @OA\Property(property="pagination", ref="#/components/schemas/Pagination")
     *         )
     *     )
     * )
     */
    public function getComptesNotArchived(Request $request)
    {
        $filters = $request->all();
        $page = (int) $request->get('page', 1);
        $limit = (int) $request->get('limit', 10);

        $comptes = $this->compteService->getAllNonArchived($filters, $page, $limit);

        return $this->successResponse(
            $comptes,
            "Liste des comptes non archives récupérée avec succès"
        );
    }

    /**
     * @OA\Post(
     *     path="/v1/comptes",
     *     tags={"Comptes"},
     *     summary="Créer un nouveau compte",
     *     description="Crée un nouveau compte bancaire avec un client. Si le client n'existe pas, il est créé automatiquement.",
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"type","soldeInitial","devise","client"},
     *             @OA\Property(property="type", type="string", enum={"cheque", "epargne"}, example="cheque"),
     *             @OA\Property(property="soldeInitial", type="number", minimum=10000, example=50000),
     *             @OA\Property(property="devise", type="string", example="FCFA"),
     *             @OA\Property(property="client", type="object",
     *                 @OA\Property(property="titulaire", type="string", example="Festo BA"),
     *                 @OA\Property(property="nci", type="string", example="1234567890987", description="Numéro NCI de 13 chiffres commençant par 1 ou 2"),
     *                 @OA\Property(property="telephone", type="string", example="781465554", description="Numéro téléphone sénégalais"),
     *                 @OA\Property(property="email", type="string", format="email", example="festobah@gmail.com"),
     *                 @OA\Property(property="adresse", type="string", example="Dakar, Sénégal")
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=201,
     *         description="Compte créé avec succès",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="success"),
     *             @OA\Property(property="http_code", type="integer", example=201),
     *             @OA\Property(property="message", type="string", example="Compte créé avec succès"),
     *             @OA\Property(property="data", type="object"),
     *             @OA\Property(property="metadata", type="object")
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Erreur de validation",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="error"),
     *             @OA\Property(property="http_code", type="integer", example=422),
     *             @OA\Property(property="message", type="string", example="Les données fournies ne sont pas valides"),
     *             @OA\Property(property="errors", type="object")
     *         )
     *     )
     * )
     */
    public function store(CompteStoreRequest $request)
    {
        try {
            $validatedData = $request->validated();

            $compte = $this->compteService->createCompte($validatedData);

            return $this->successResponse(
                $compte->load('client'),
                "Compte créé avec succès",
                Response::HTTP_CREATED
            );
        } catch (\Throwable $e) {
            return $this->errorResponse(
                "Erreur lors de la création du compte",
                422
            );
        }
    }

    /**
     * @OA\Get(
     *     path="/v1/comptes/{id}",
     *     tags={"Comptes"},
     *     summary="Affiche un compte spécifique",
     *     description="Récupère les détails d'un compte par son ID",
     *     security={{"bearerAuth":{}}},
     *     @OA\Parameter(
     *         name="id",
     *         in="path",
     *         description="ID du compte",
     *         required=true,
     *         @OA\Schema(type="string")
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Compte récupéré avec succès",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="success"),
     *             @OA\Property(property="http_code", type="integer", example=200),
     *             @OA\Property(property="message", type="string", example="Compte récupéré avec succès"),
     *             @OA\Property(property="data", type="object")
     *         )
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="Compte introuvable",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="error"),
     *             @OA\Property(property="http_code", type="integer", example=404),
     *             @OA\Property(property="message", type="string", example="Compte introuvable")
     *         )
     *     )
     * )
     */
    public function show(int|string $id)
    {
        $compte = $this->compteService->getById($id);

        if (!$compte) {
            return $this->errorResponse("Compte introuvable", 404);
        }

        return $this->successResponse(
            new CompteResource($compte, $this->clientService),
            "Compte récupéré avec succès"
        );
    }

    /**
     * @OA\Put(
     *     path="/v1/comptes/{id}",
     *     tags={"Comptes"},
     *     summary="Mettre à jour un compte",
     *     description="Met à jour les informations d'un compte bancaire existant",
     *     security={{"bearerAuth":{}}},
     *     @OA\Parameter(
     *         name="id",
     *         in="path",
     *         description="ID du compte à mettre à jour",
     *         required=true,
     *         @OA\Schema(type="string")
     *     ),
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             @OA\Property(property="type", type="string", enum={"cheque", "epargne", "courant"}, example="epargne"),
     *             @OA\Property(property="solde_initial", type="number", minimum=10000, example=50000),
     *             @OA\Property(property="devise", type="string", example="FCFA"),
     *             @OA\Property(property="statut", type="string", enum={"actif", "inactif", "bloque", "ferme"}, example="actif"),
     *             @OA\Property(property="metadata", type="object", example={"notes": "Mise à jour du solde"})
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Compte mis à jour avec succès",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="success"),
     *             @OA\Property(property="http_code", type="integer", example=200),
     *             @OA\Property(property="message", type="string", example="Compte mis à jour avec succès"),
     *             @OA\Property(property="data", type="object")
     *         )
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="Compte introuvable",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="error"),
     *             @OA\Property(property="http_code", type="integer", example=404),
     *             @OA\Property(property="message", type="string", example="Compte introuvable")
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Erreur de validation",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="error"),
     *             @OA\Property(property="http_code", type="integer", example=422),
     *             @OA\Property(property="message", type="string", example="Les données fournies ne sont pas valides"),
     *             @OA\Property(property="errors", type="object")
     *         )
     *     )
     * )
     */
    public function update(CompteUpdateRequest $request, int|string $id)
    {
        try {
            $validatedData = $request->validated();

            $compte = $this->compteService->updateCompte($id, $validatedData);

            return $this->successResponse(
                new CompteResource($compte, $this->clientService),
                "Compte mis à jour avec succès"
            );
        } catch (\Throwable $e) {
            return $this->errorResponse(
                $e->getMessage(),
                422
            );
        }
    }

    /**
     * @OA\Post(
     *     path="/v1/comptes/{id}/bloquer",
     *     tags={"Comptes"},
     *     summary="Bloquer un compte épargne",
     *     description="Bloque un compte épargne actif pour une durée déterminée en mois avec un motif spécifique",
     *     security={{"bearerAuth":{}}},
     *     @OA\Parameter(
     *         name="id",
     *         in="path",
     *         description="ID du compte épargne à bloquer",
     *         required=true,
     *         @OA\Schema(type="string")
     *     ),
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"motif","duree","unite"},
     *             @OA\Property(property="motif", type="string", example="Activité suspecte détectée"),
     *             @OA\Property(property="duree", type="integer", minimum=1, example=3),
     *             @OA\Property(property="unite", type="string", enum={"mois"}, example="mois")
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Compte bloqué avec succès",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="success"),
     *             @OA\Property(property="http_code", type="integer", example=200),
     *             @OA\Property(property="message", type="string", example="Compte bloqué avec succès"),
     *             @OA\Property(property="data", type="object")
     *         )
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="Compte introuvable",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="error"),
     *             @OA\Property(property="http_code", type="integer", example=404),
     *             @OA\Property(property="message", type="string", example="Compte introuvable")
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Erreur de validation ou contraintes métier non respectées",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="error"),
     *             @OA\Property(property="http_code", type="integer", example=422),
     *             @OA\Property(property="message", type="string", example="Seuls les comptes épargne actifs peuvent être bloqués"),
     *             @OA\Property(property="errors", type="object")
     *         )
     *     )
     * )
     */
    public function bloquer(CompteBloquerRequest $request, int|string $id)
    {
        try {
            $validatedData = $request->validated();

            $compte = $this->compteService->bloquerCompte($id, $validatedData);

            return $this->successResponse(
                new CompteResource($compte, $this->clientService),
                "Compte bloqué avec succès"
            );
        } catch (\Throwable $e) {
            return $this->errorResponse(
                $e->getMessage(),
                422
            );
        }
    }

    /**
     * @OA\Post(
     *     path="/v1/comptes/{id}/debloquer",
     *     tags={"Comptes"},
     *     summary="Débloquer un compte manuellement",
     *     description="Débloque un compte bloqué sur demande du client",
     *     security={{"bearerAuth":{}}},
     *     @OA\Parameter(
     *         name="id",
     *         in="path",
     *         description="ID du compte à débloquer",
     *         required=true,
     *         @OA\Schema(type="string")
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Compte débloqué avec succès",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="success"),
     *             @OA\Property(property="http_code", type="integer", example=200),
     *             @OA\Property(property="message", type="string", example="Compte débloqué avec succès"),
     *             @OA\Property(property="data", type="object")
     *         )
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="Compte introuvable",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="error"),
     *             @OA\Property(property="http_code", type="integer", example=404),
     *             @OA\Property(property="message", type="string", example="Compte introuvable")
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Erreur de validation",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="error"),
     *             @OA\Property(property="http_code", type="integer", example=422),
     *             @OA\Property(property="message", type="string", example="Le compte n'est pas bloqué")
     *         )
     *     )
     * )
     */
    public function debloquer(int|string $id)
    {
        try {
            $compte = $this->compteService->debloquerCompteManuellement($id);

            return $this->successResponse(
                new CompteResource($compte, $this->clientService),
                "Compte débloqué avec succès"
            );
        } catch (\Throwable $e) {
            return $this->errorResponse(
                $e->getMessage(),
                422
            );
        }
    }

    /**
     * @OA\Post(
     *     path="/v1/comptes/debloquer-expiration",
     *     tags={"Comptes"},
     *     summary="Débloquer automatiquement les comptes expirés",
     *     description="Débloque automatiquement tous les comptes dont la période de blocage est arrivée à expiration",
     *     security={{"bearerAuth":{}}},
     *     @OA\Response(
     *         response=200,
     *         description="Comptes débloqués avec succès",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="string", example="success"),
     *             @OA\Property(property="http_code", type="integer", example=200),
     *             @OA\Property(property="message", type="string", example="3 comptes débloqués automatiquement"),
     *             @OA\Property(property="data", type="array", @OA\Items(type="object"))
     *         )
     *     )
     * )
     */
    public function debloquerExpiration()
    {
        try {
            $comptesDebloques = $this->compteService->debloquerComptesExpires();

            return $this->successResponse(
                collect($comptesDebloques)->map(function ($compte) {
                    return new CompteResource($compte, $this->clientService);
                }),
                count($comptesDebloques) . " comptes débloqués automatiquement"
            );
        } catch (\Throwable $e) {
            return $this->errorResponse(
                $e->getMessage(),
                422
            );
        }
    }
}