# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion des Comptes",
        "description": "API pour la gestion des comptes bancaires",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://jeeri.onrender.com/tinkin",
            "description": "Serveur de production"
        },
        {
            "url": "http://localhost:8000/tinkin",
            "description": "Serveur local"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/v1/comptes/non-archives": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Liste des comptes non archivés",
                "description": "Récupère tous les comptes non archivés avec possibilité de filtrage",
                "operationId": "18aeee1839e78c53da6132a1e353d58b",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "cheque",
                                "epargne",
                                "courant"
                            ]
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "id"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des comptes non archivés récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "http_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des comptes non archives récupérée avec succès"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Compte"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/comptes/{id}/bloquer": {
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Bloquer un compte épargne",
                "description": "Bloque un compte épargne actif pour une durée déterminée en mois avec un motif spécifique",
                "operationId": "5941080a253593b87c28b5df0d5fdaf8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du compte épargne à bloquer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "motif",
                                    "duree",
                                    "unite"
                                ],
                                "properties": {
                                    "motif": {
                                        "type": "string",
                                        "example": "Activité suspecte détectée"
                                    },
                                    "duree": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 3
                                    },
                                    "unite": {
                                        "type": "string",
                                        "enum": [
                                            "mois"
                                        ],
                                        "example": "mois"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compte bloqué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "http_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte bloqué avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "http_code": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte introuvable"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation ou contraintes métier non respectées",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "http_code": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Seuls les comptes épargne actifs peuvent être bloqués"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/v1/comptes/debloquer-expiration": {
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Débloquer automatiquement les comptes expirés",
                "description": "Débloque automatiquement tous les comptes dont la période de blocage est arrivée à expiration",
                "operationId": "cbfc4f123f588a02aadaf872fc65dc86",
                "responses": {
                    "200": {
                        "description": "Comptes débloqués avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "http_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "3 comptes débloqués automatiquement"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Pagination": {
                "title": "Pagination",
                "description": "Informations de pagination",
                "properties": {
                    "currentPage": {
                        "type": "integer"
                    },
                    "totalPages": {
                        "type": "integer"
                    },
                    "totalItems": {
                        "type": "integer"
                    },
                    "itemsPerPage": {
                        "type": "integer"
                    },
                    "hasNext": {
                        "type": "boolean"
                    },
                    "hasPrevious": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            }
        }
    },
...
}